requirements.txt
networkx
scipy
scikit-learn
tqdm
matplotlib
git+https://github.com/fegerar/shapegraphs.git
lightning
ogb
torch_geometric
wandb
pyyaml
